% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_api_connection.R
\name{check_api_status}
\alias{check_api_status}
\title{Check OpenNEM API status}
\usage{
check_api_status(suppress_messages = TRUE)
}
\arguments{
\item{suppress_messages}{suppress chromote messages}
}
\value{
status of the OpenNEM API
}
\description{
Check OpenNEM API status
}
\examples{
check_api_status()

}
