% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_api_connection.R
\name{check_api_version}
\alias{check_api_version}
\title{Check API version number}
\usage{
check_api_version(suppress_messages = TRUE)
}
\arguments{
\item{suppress_messages}{suppress chromote messages}
}
\value{
[string] API version number
}
\description{
Check API version number
}
\examples{
check_api_version()

}
