% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{format_datetime}
\alias{format_datetime}
\title{Helper function to format datetime for API}
\usage{
format_datetime(dt)
}
\arguments{
\item{dt}{Character or POSIXct datetime}
}
\value{
Character in ISO 8601 format without timezone
}
\description{
Helper function to format datetime for API
}
\keyword{internal}
