% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_fueltech_list.R
\name{get_fueltech_list}
\alias{get_fueltech_list}
\title{Get list of fueltechs available in OpenNEM API}
\usage{
get_fueltech_list()
}
\value{
Dataframe of fueltechs
}
\description{
Get list of fueltechs available in OpenNEM API
}
\examples{
get_fueltech_list()

}
