% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_interval_list.R
\name{get_interval_list}
\alias{get_interval_list}
\title{Get list of intervals available in OpenNEM API}
\usage{
get_interval_list()
}
\value{
Dataframe of intervals
}
\description{
Get list of intervals available in OpenNEM API
}
\examples{
get_interval_list()
}
