% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_network_list.R
\name{get_network_list}
\alias{get_network_list}
\title{Get list of networks available in OpenNEM API}
\usage{
get_network_list()
}
\value{
Dataframe of networks and network regions
}
\description{
Get list of networks available in OpenNEM API
}
\examples{
get_network_list()

}
