% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_period_list.R
\name{get_period_list}
\alias{get_period_list}
\title{Get list of periods available in OpenNEM API}
\usage{
get_period_list()
}
\value{
Dataframe of periods
}
\description{
Get list of periods available in OpenNEM API
}
\examples{
get_period_list()
}
