% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_power_by_station.R
\name{get_power_by_station}
\alias{get_power_by_station}
\title{Get historical power by station data}
\usage{
get_power_by_station(
  network_code,
  station_code,
  interval = "30m",
  period = "7d"
)
}
\description{
Get historical power by station data
}
