% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_station_list.R
\name{get_station_list}
\alias{get_station_list}
\title{Get list of stations available in OpenNEM}
\usage{
get_station_list()
}
\value{
Dataframe of stations and facilities
}
\description{
Get list of stations available in OpenNEM
}
\examples{
df <- get_station_list()
}
