% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{oe_network_interconnectors}
\alias{oe_network_interconnectors}
\title{OpenElectricity Network Interconnectors}
\format{
A tibble with 4 rows and 3 columns:
\describe{
  \item{interconnector}{Interconnector code showing connected regions (e.g., "NSW1-QLD1")}
  \item{interconnector_description}{Descriptive name of the interconnector}
  \item{network_name}{Parent network (currently only NEM has interconnectors)}
}
}
\source{
OpenElectricity API v4
}
\usage{
oe_network_interconnectors
}
\description{
Reference dataset containing information about grid interconnections between regions.
}
\details{
Interconnectors allow electricity to flow between regional grids:
- NSW1-QLD1: Queensland to New South Wales
- VIC1-NSW1: Victoria to New South Wales
- VIC1-SA1: Victoria to South Australia
- VIC1-TAS1: Victoria to Tasmania (Basslink submarine cable)
}
\examples{
# View all interconnectors
oe_network_interconnectors

# Find interconnectors involving Victoria
subset(oe_network_interconnectors, grepl("VIC1", interconnector))

}
\keyword{datasets}
