% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{oe_network_regions}
\alias{oe_network_regions}
\title{OpenElectricity Network Regions}
\format{
A tibble with 7 rows and 3 columns:
\describe{
  \item{region}{Region code (NSW1, QLD1, SA1, TAS1, VIC1, SNOWY1, WEM)}
  \item{region_name}{Full descriptive name of the region}
  \item{network_name}{Parent network identifier (NEM or WEM)}
}
}
\source{
OpenElectricity API v4
}
\usage{
oe_network_regions
}
\description{
Reference dataset containing region codes and names for each electricity network.
}
\details{
NEM regions include:
- NSW1: New South Wales (including ACT)
- QLD1: Queensland
- SA1: South Australia
- TAS1: Tasmania
- VIC1: Victoria
- SNOWY1: Snowy Mountains region

WEM has a single region also called WEM (Western Australia).
}
\examples{
# View all regions
oe_network_regions

# Get NEM regions only
subset(oe_network_regions, network_name == "NEM")

}
\keyword{datasets}
