% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{oe_primary_groupings}
\alias{oe_primary_groupings}
\title{OpenElectricity Primary Groupings}
\format{
A tibble with 2 rows and 2 columns:
\describe{
  \item{primary_grouping}{Grouping identifier code}
  \item{grouping_desc}{Description of the grouping}
}
}
\source{
\url{https://docs.openelectricity.org.au/api-reference/data/get-network-data}
}
\usage{
oe_primary_groupings
}
\description{
Reference dataset containing valid primary grouping options for network data
API queries.
}
\keyword{datasets}
