% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{oe_secondary_groupings}
\alias{oe_secondary_groupings}
\title{OpenElectricity Secondary Groupings}
\format{
A tibble with 4 rows and 2 columns:
\describe{
  \item{secondary_grouping}{Grouping identifier code}
  \item{grouping_desc}{Description of the grouping}
}
}
\source{
\url{https://docs.openelectricity.org.au/api-reference/data/get-network-data}
}
\usage{
oe_secondary_groupings
}
\description{
Reference dataset containing valid secondary grouping options for network data
API queries. Secondary groupings add a second dimension on top of the primary
grouping.
}
\keyword{datasets}
