% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parse_facilities_data.R
\name{parse_facilities_data}
\alias{parse_facilities_data}
\title{Helper function to parse facilities API response data into a data frame}
\usage{
parse_facilities_data(data)
}
\arguments{
\item{data}{List from API response}
}
\value{
Data frame with fully unnested location and units
}
\description{
Helper function to parse facilities API response data into a data frame
}
\keyword{internal}
