% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zzz.R
\name{repeat_df_rows}
\alias{repeat_df_rows}
\title{Repeat dataframe n times}
\usage{
repeat_df_rows(df, n)
}
\arguments{
\item{df}{A dataframe object}

\item{n}{The number of times to repeat the dataframe rows}
}
\value{
A dataframe with original rows repeated n times
}
\description{
Repeat dataframe n times
}
