% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/response_code_message.R
\name{response_code_message}
\alias{response_code_message}
\title{HTML response code handler}
\usage{
response_code_message(code)
}
\arguments{
\item{code}{numeric response code from GET request}
}
\value{
HTML response code message for relevant code
}
\description{
HTML response code handler
}
\examples{
\dontrun{
response_code_message(200)
}

}
