% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/error_codes.R
\name{response_undefined}
\alias{response_undefined}
\title{Title}
\usage{
response_undefined(response_code)
}
\arguments{
\item{response_code}{}
}
\value{
Message against defined and undefined html response codes
}
\description{
Title
}
\examples{
response_undefined()
}
