% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parse_network_data.R
\name{parse_network_data}
\alias{parse_network_data}
\title{Helper function to parse API response data into a data frame}
\usage{
parse_network_data(data)
}
\arguments{
\item{data}{List from API response}
}
\value{
Data frame
}
\description{
Helper function to parse API response data into a data frame
}
\keyword{internal}
