% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/response_code_message.R
\name{response_code_message}
\alias{response_code_message}
\title{HTML response code handler}
\usage{
response_code_message(code)
}
\value{
Return defined message for each html response code
}
\description{
HTML response code handler
}
